+++ examples/various_settings_task.rs
@@ -1,28 +1,24 @@
-use chrono::prelude::*;
-use planif::enums::TaskCreationFlags;
-<<<<<<< HEAD
-use planif::schedule_builder::{Action, ComRuntime, ScheduleBuilder};
-=======
-use planif::schedule_builder::{Action, ScheduleBuilder};
->>>>>>> origin/main
-use planif::settings::Settings;
+// use chrono::prelude::*;
+// use planif::enums::TaskCreationFlags;
+// use planif::schedule_builder::{Action, ComRuntime, ScheduleBuilder};
+// use planif::settings::Settings;
 
-fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let com = ComRuntime::new()?;
-    let sb = ScheduleBuilder::new(&com).unwrap();
-    let mut settings = Settings::new();
-    settings.run_only_if_idle = Some(true);
+// fn main() -> Result<(), Box<dyn std::error::Error>> {
+//     let com = ComRuntime::new()?;
+//     let sb = ScheduleBuilder::new(&com).unwrap();
+//     let mut settings = Settings::new();
+//     settings.run_only_if_idle = Some(true);
 
-    sb.create_daily()
-        .author("Matt")?
-        .description("Test Trigger")?
-        .settings(settings)?
-        .trigger("test_trigger", true)?
-        .days_interval(1)?
-        .action(Action::new("test", "notepad.exe", "", ""))?
-        .start_boundary(&Local::now().to_rfc3339())?
-        .build()?
-        .register("TaskName", TaskCreationFlags::CreateOrUpdate as i32)?;
+//     sb.create_daily()
+//         .author("Matt")?
+//         .description("Test Trigger")?
+//         .settings(settings)?
+//         .trigger("test_trigger", true)?
+//         .days_interval(1)?
+//         .action(Action::new("test", "notepad.exe", "", ""))?
+//         .start_boundary(&Local::now().to_rfc3339())?
+//         .build()?
+//         .register("TaskName", TaskCreationFlags::CreateOrUpdate as i32)?;
 
-    Ok(())
-}
+//     Ok(())
+// }