+++ src/schedule_builder.rs
@@ -1,10 +1,14 @@
 use crate::{
     enums::{DayOfMonth, DayOfWeek, Month, WeekOfMonth},
     error::{InvalidOperationError, RequiredPropertyError},
-    schedule::Schedule,
-    settings::{Duration, PrincipalSettings, Settings},
+    schedule::{Schedule, Unregistered},
+    settings::{Duration, PrincipalSettings, Settings}, task_scheduler::ComRuntime,
     com::ComRuntime,
 };
+
+use windows::core::ComInterface;
+use windows::core::BSTR;
+use windows::Win32::Foundation::VARIANT_BOOL;
 use windows::Win32::System::Com::{
     CoCreateInstance, CLSCTX_ALL, VARIANT,
 };
@@ -14,8 +18,7 @@ use windows::Win32::System::TaskScheduler::{
     IAction, IActionCollection, IBootTrigger, IDailyTrigger, IEventTrigger, IExecAction,
     IIdleTrigger, ILogonTrigger, IMonthlyDOWTrigger, IMonthlyTrigger, INetworkSettings, IPrincipal,
     IRegistrationInfo, IRegistrationTrigger, IRepetitionPattern, ITaskDefinition, ITaskFolder,
-    ITaskService, ITaskSettings, ITimeTrigger, ITrigger, ITriggerCollection, IWeeklyTrigger,
-    TaskScheduler, TASK_ACTION_EXEC, TASK_LOGON_TYPE, TASK_RUNLEVEL_TYPE, TASK_TRIGGER_BOOT,
+    ITaskService, ITaskSettings, ITimeTrigger, ITrigger, ITriggerCollection, IWeeklyTrigger, TASK_ACTION_EXEC, TASK_LOGON_TYPE, TASK_RUNLEVEL_TYPE, TASK_TRIGGER_BOOT,
     TASK_TRIGGER_DAILY, TASK_TRIGGER_EVENT, TASK_TRIGGER_IDLE, TASK_TRIGGER_LOGON,
     TASK_TRIGGER_MONTHLY, TASK_TRIGGER_MONTHLYDOW, TASK_TRIGGER_REGISTRATION, TASK_TRIGGER_TIME,
     TASK_TRIGGER_WEEKLY,
@@ -51,7 +54,8 @@ pub struct Weekly {}
 pub struct ScheduleBuilder<Frequency = Base> {
     pub(crate) frequency: std::marker::PhantomData<Frequency>,
     pub(crate) schedule: Schedule,
-    com: ComRuntime,
+    com_runtime: ComRuntime,
+    task_service: ITaskService
 }
 
 impl ScheduleBuilder<Base> {
@@ -64,17 +68,20 @@ impl ScheduleBuilder<Base> {
     /// let com = ComRuntime::new()?;
     /// let builder: ScheduleBuilder<Base> = ScheduleBuilder::new(&com).unwrap();
     /// ```
-    pub fn new(com: &ComRuntime) -> Result<Self, Box<dyn std::error::Error>> {
+    pub fn new(com_runtime: &ComRuntime, task_service: &ITaskService) -> Result<Self, Box<dyn std::error::Error>> {
         unsafe {
-            let sb_com = com.clone();
+            let sb_com_rt = com_runtime.clone();
+            let sb_ts = task_service.clone();
 
-            let task_service: ITaskService = CoCreateInstance(&TaskScheduler, None, CLSCTX_ALL)?;
-            task_service.Connect(
-                VARIANT::default(),
-                VARIANT::default(),
-                VARIANT::default(),
-                VARIANT::default(),
-            )?;
+            let sch_com_rt = com_runtime.clone();
+
+            // let task_service: ITaskService = CoCreateInstance(&TaskScheduler, None, CLSCTX_ALL)?;
+            // task_service.Connect(
+            //     VARIANT::default(),
+            //     VARIANT::default(),
+            //     VARIANT::default(),
+            //     VARIANT::default(),
+            // )?;
 
             let task_folder: ITaskFolder = task_service.GetFolder(&BSTR::from("\\"))?;
             let task_definition: ITaskDefinition = task_service.NewTask(0)?;
@@ -84,18 +91,21 @@ impl ScheduleBuilder<Base> {
             let settings: ITaskSettings = task_definition.Settings()?;
 
             Ok(Self {
-                com: sb_com,
+                com_runtime: sb_com_rt,
                 frequency: std::marker::PhantomData::<Base>,
+                task_service: sb_ts,
                 schedule: Schedule {
+                    kind: std::marker::PhantomData::<Unregistered>,
                     task_folder,
                     actions,
                     force_start_boundary: false,
                     registration_info,
                     settings,
-                    task_service,
+                    // task_service,
                     task_definition,
                     trigger: None,
                     triggers,
+                    com_runtime: sch_com_rt
                 },
             })
         }
@@ -114,9 +124,10 @@ impl ScheduleBuilder<Base> {
     /// ```
     pub fn create_boot(self) -> ScheduleBuilder<Boot> {
         ScheduleBuilder::<Boot> {
-            com: self.com,
+            com_runtime: self.com_runtime,
             frequency: std::marker::PhantomData::<Boot>,
             schedule: self.schedule,
+            task_service: self.task_service
         }
     }
 
@@ -134,9 +145,10 @@ impl ScheduleBuilder<Base> {
     pub fn create_daily(mut self) -> ScheduleBuilder<Daily> {
         self.schedule.force_start_boundary = true;
         ScheduleBuilder::<Daily> {
-            com: self.com,
+            com_runtime: self.com_runtime,
             frequency: std::marker::PhantomData::<Daily>,
             schedule: self.schedule,
+            task_service: self.task_service
         }
     }
 
@@ -154,9 +166,10 @@ impl ScheduleBuilder<Base> {
     pub fn create_event(mut self) -> ScheduleBuilder<Event> {
         self.schedule.force_start_boundary = true;
         ScheduleBuilder::<Event> {
-            com: self.com,
+            com_runtime: self.com_runtime,
             frequency: std::marker::PhantomData::<Event>,
             schedule: self.schedule,
+            task_service: self.task_service
         }
     }
 
@@ -173,9 +186,10 @@ impl ScheduleBuilder<Base> {
     /// ```
     pub fn create_idle(self) -> ScheduleBuilder<Idle> {
         ScheduleBuilder::<Idle> {
-            com: self.com,
+            com_runtime: self.com_runtime,
             frequency: std::marker::PhantomData::<Idle>,
             schedule: self.schedule,
+            task_service: self.task_service
         }
     }
 
@@ -192,9 +206,10 @@ impl ScheduleBuilder<Base> {
     /// ```
     pub fn create_logon(self) -> ScheduleBuilder<Logon> {
         ScheduleBuilder::<Logon> {
-            com: self.com,
+            com_runtime: self.com_runtime,
             frequency: std::marker::PhantomData::<Logon>,
             schedule: self.schedule,
+            task_service: self.task_service
         }
     }
 
@@ -210,9 +225,10 @@ impl ScheduleBuilder<Base> {
     /// ```
     pub fn create_monthly(self) -> ScheduleBuilder<Monthly> {
         ScheduleBuilder::<Monthly> {
-            com: self.com,
+            com_runtime: self.com_runtime,
             frequency: std::marker::PhantomData::<Monthly>,
             schedule: self.schedule,
+            task_service: self.task_service
         }
     }
 
@@ -229,9 +245,10 @@ impl ScheduleBuilder<Base> {
     pub fn create_monthly_dow(mut self) -> ScheduleBuilder<MonthlyDOW> {
         self.schedule.force_start_boundary = true;
         ScheduleBuilder::<MonthlyDOW> {
-            com: self.com,
+            com_runtime: self.com_runtime,
             frequency: std::marker::PhantomData::<MonthlyDOW>,
             schedule: self.schedule,
+            task_service: self.task_service
         }
     }
 
@@ -247,9 +264,10 @@ impl ScheduleBuilder<Base> {
     /// ```
     pub fn create_registration(self) -> ScheduleBuilder<Registration> {
         ScheduleBuilder::<Registration> {
-            com: self.com,
+            com_runtime: self.com_runtime,
             frequency: std::marker::PhantomData::<Registration>,
             schedule: self.schedule,
+            task_service: self.task_service
         }
     }
 
@@ -267,9 +285,10 @@ impl ScheduleBuilder<Base> {
     pub fn create_time(mut self) -> ScheduleBuilder<Time> {
         self.schedule.force_start_boundary = true;
         ScheduleBuilder::<Time> {
-            com: self.com,
+            com_runtime: self.com_runtime,
             frequency: std::marker::PhantomData::<Time>,
             schedule: self.schedule,
+            task_service: self.task_service
         }
     }
 
@@ -286,9 +305,10 @@ impl ScheduleBuilder<Base> {
     /// ```
     pub fn create_weekly(self) -> ScheduleBuilder<Weekly> {
         ScheduleBuilder::<Weekly> {
-            com: self.com,
+            com_runtime: self.com_runtime,
             frequency: std::marker::PhantomData::<Weekly>,
             schedule: self.schedule,
+            task_service: self.task_service
         }
     }
 }
@@ -313,10 +333,9 @@ impl<Frequency> ScheduleBuilder<Frequency> {
         unsafe {
             // check if folder exists, if not make it
             self.schedule.task_folder =
-                match self.schedule.task_service.GetFolder(&BSTR::from(folder)) {
+                match self.task_service.GetFolder(&BSTR::from(folder)) {
                     Ok(x) => x,
                     Err(_) => self
-                        .schedule
                         .task_service
                         .GetFolder(&BSTR::from("\\"))
                         .unwrap()