+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.2]
+
 ### Added
 * Add `Duration` struct to ease usage of delays and other methods using string duration formats
 * Add Default impl for `Duration`
@@ -27,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Moved `ComRuntime` to `planif::com` module
 
 ## [0.2.1]
+
 ### Updated
 * Update documentation, add deny missing_docs